requirements.txt
astroid==1.6.1
Django==1.11.10
django-crispy-forms==1.6.1
django-extensions==1.9.9
django-filter==1.1.0
django-model-utils==3.1.1
django-render-block==0.5
djangorestframework==3.5.3
djangorestframework-jwt==1.11.0
html5lib==1.0b8
httplib2==0.10.3
isort==4.2.15
jsonfield==2.0.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
PyJWT==1.6.1
pylint==1.8.2
PyPDF2==1.26.0
pytz==2017.3
reportlab==3.4.0
six==1.11.0
typing==3.6.4
webencodings==0.5.1
wrapt==1.10.11
xhtml2pdf==0.1b2
sendgrid==3.0.0
numpy==1.14.5



